Remove unnecessary grapl_common dependence in analyzerlib
None of the code in `grapl_analyzerlib` actually calls anything in
`grapl_common`, so we can the bits that claim otherwise.

This revealed the ways that our current build system obscures some of
the real dependencies, however. Since everything basically uses the
virtualenv of the analyzerlib, anything that was included there could
be used to silently satisfy dependencies. By removing `grapl-common`
from `analyzerlib`, we revealed three other packages that were
satisfying their dependency indirectly:

* analyzer-executor
* engagement-creator
* grapl-tests-common

Since our current Python monorepo and build system aren't really
organized to deal with the specific needs of interdependent Python
packages in a clean way, we are going to resort to a bit of a
temporary hack to deal with these dependencies, pending a more
extensive reorganization of the code.

Basically, we'll copy `grapl-common` into the build images that need
it and `pip install` it into the virtual environment.

This isn't ideal, since it doesn't capture the dependency in an
obvious or Python-standard manner. It does, however, present a minimal
solution that can be extended to other services as similar
dependency issues are uncovered. This in turn will allow us to
incrementally improve and tighten up our Python code structure. The
long-term goal is to not require such container build-time
shenanigans.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>